Challenge day 1

+++ Project1.java
@@ -5,12 +5,12 @@ public class Project1 {
         Scanner sc = new Scanner(System.in);
 
         System.out.println("Digite o primeiro número: ");
-        int number1 = sc.nextInt();
+        double number1 = sc.nextDouble();
         
         System.out.println("Digite o segundo número: ");
-        int number2 = sc.nextInt();
+        double number2 = sc.nextDouble();
 
-        int sum = number1 + number2;
+        double sum = number1 + number2;
         System.out.println("A soma é : " + sum);
 
         sc.close();
